Shifted TokenActivity logic to ViewModel to use Coroutines and LiveData
