+++ app/src/main/java/crux/bphc/cms/models/SingleLiveEvent.kt
@@ -1,0 +1,14 @@
+package crux.bphc.cms.models
+
+class SingleLiveEvent<T>(private val content: T) {
+    private var hasBeenHandled = false
+
+    fun getContentIfNotHandled(): T? {
+        return if (hasBeenHandled) {
+            null
+        } else {
+            hasBeenHandled = true
+            content
+        }
+    }
+}
